# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "trao interview prep kit",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/kits/{kit_id}/{collection}": {
      "post": {
        "summary": "Create Item",
        "operationId": "create_item_api_kits__kit_id___collection__post",
        "parameters": [
          {
            "name": "kit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kit Id"
            }
          },
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/QuestionOut"
                    },
                    {
                      "$ref": "#/components/schemas/FlashcardOut"
                    },
                    {
                      "$ref": "#/components/schemas/RequirementOut"
                    },
                    {
                      "$ref": "#/components/schemas/BriefOut"
                    }
                  ],
                  "title": "Response Create Item Api Kits  Kit Id   Collection  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kits/{kit_id}/{collection}/{item_id}": {
      "patch": {
        "summary": "Patch Item",
        "operationId": "patch_item_api_kits__kit_id___collection___item_id__patch",
        "parameters": [
          {
            "name": "kit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kit Id"
            }
          },
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/QuestionOut"
                    },
                    {
                      "$ref": "#/components/schemas/FlashcardOut"
                    },
                    {
                      "$ref": "#/components/schemas/RequirementOut"
                    },
                    {
                      "$ref": "#/components/schemas/BriefOut"
                    }
                  ],
                  "title": "Response Patch Item Api Kits  Kit Id   Collection   Item Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Item",
        "operationId": "delete_item_api_kits__kit_id___collection___item_id__delete",
        "parameters": [
          {
            "name": "kit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kit Id"
            }
          },
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kits/{kit_id}/questions/reorder": {
      "post": {
        "summary": "Reorder",
        "operationId": "reorder_api_kits__kit_id__questions_reorder_post",
        "parameters": [
          {
            "name": "kit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kit Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BriefOut": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "what_they_do": {
            "type": "string",
            "title": "What They Do"
          },
          "hiring_process": {
            "type": "string",
            "title": "Hiring Process"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "summary",
          "what_they_do",
          "hiring_process",
          "sources"
        ],
        "title": "BriefOut"
      },
...
      "DeleteItemOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "flagged": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flagged"
          }
        },
        "type": "object",
        "title": "DeleteItemOut"
      },
...
      "FlashcardOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "front": {
            "type": "string",
            "title": "Front"
          },
          "back": {
            "type": "string",
            "title": "Back"
          },
          "requirement_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requirement Ids"
          },
          "_meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItemMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "front",
          "back",
          "requirement_ids"
        ],
        "title": "FlashcardOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemMeta": {
        "properties": {
          "origin": {
            "type": "string",
            "title": "Origin",
            "default": "generated"
          },
          "edited": {
            "type": "boolean",
            "title": "Edited",
            "default": false
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned",
            "default": false
          },
          "rev": {
            "type": "integer",
            "title": "Rev",
            "default": 1
          },
          "order": {
            "type": "string",
            "title": "Order",
            "default": ""
          },
          "gen_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gen Run"
          }
        },
        "type": "object",
        "title": "ItemMeta"
      },
...
      "QuestionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "requirement_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requirement Ids"
          },
          "category": {
            "type": "string",
            "enum": [
              "technical",
              "behavioural",
              "system-design",
              "company-fit"
            ],
            "title": "Category"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "answer_outline": {
            "type": "string",
            "title": "Answer Outline"
          },
          "difficulty": {
            "type": "integer",
            "title": "Difficulty"
          },
          "outline_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Outline Points"
          },
          "_meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItemMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "requirement_ids",
          "category",
          "prompt",
          "answer_outline",
          "difficulty",
          "outline_points"
        ],
        "title": "QuestionOut"
      },
...
      "RequirementOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "kind": {
            "type": "string",
            "enum": [
              "technical",
              "behavioural",
              "domain"
            ],
            "title": "Kind"
          },
          "priority": {
            "type": "string",
            "enum": [
              "must",
              "nice"
            ],
            "title": "Priority"
          },
          "_meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItemMeta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "text",
          "kind",
          "priority"
        ],
        "title": "RequirementOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}